fix: add deposited status
