+++ src/types/api.ts
@@ -1,20 +1,4 @@
-export const VALIDATOR_STATUS = [
-  'pending_initialized',
-  'pending_queued',
-  'active_ongoing',
-  'active_online',
-  'active_offline',
-  'active_exiting',
-  'active_slashed',
-  'exited_unslashed',
-  'exited',
-  'exited_slashed',
-  'withdrawal_possible',
-  'withdrawal_done',
-] as const
-
-export type ValidatorStatus =
-  typeof VALIDATOR_STATUS[number]
+import { ValidatorStatus } from "./beacon";
 
 export interface APIValidatorDetailsResponse {
     activationeligibilityepoch: number;
